# documentation/api-docs-auth-service.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8081",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/test/kyc/passport/verification": {
      "post": {
        "tags": [
          "test-kyc-controller"
        ],
        "operationId": "uploadPassportData",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadPassportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PassportVerificationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/register": {
      "post": {
        "tags": [
          "Authorization and authentication"
        ],
        "description": "Регистрация пользователя",
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/refresh": {
      "post": {
        "tags": [
          "Authorization and authentication"
        ],
        "description": "Обновить токены",
        "operationId": "refreshTokens",
        "parameters": [
          {
            "name": "refresh",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": [
          "Authorization and authentication"
        ],
        "description": "Выход из системы (логаут)",
        "operationId": "logout",
        "parameters": [
          {
            "name": "refresh",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": [
          "Authorization and authentication"
        ],
        "description": "Вход в систему (логин)",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/verification/kyc/cancel/{userId}": {
      "post": {
        "tags": [
          "Admin Verification management"
        ],
        "description": "Отклонить верификацию аккаунта пользователя",
        "operationId": "cancelVerification",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/admin/verification/kyc/approve/{userId}": {
      "post": {
        "tags": [
          "Admin Verification management"
        ],
        "description": "Подтвердить верификацию аккаунта пользователя",
        "operationId": "approveVerification",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/users/{userId}/email": {
      "get": {
        "tags": [
          "Получить email пользователя"
        ],
        "operationId": "getUserEmail",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/view/kyc/verified": {
      "get": {
        "tags": [
          "Admin View Management"
        ],
        "description": "Получить все верифицированные заявки",
        "operationId": "getCheckedVerifications",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationViewByAdminResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/view/kyc/pending": {
      "get": {
        "tags": [
          "Admin View Management"
        ],
        "description": "Получить все заявки пользователей на верификацию аккаунта",
        "operationId": "getPendingVerifications",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationViewByAdminResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/view/kyc/not-verified": {
      "get": {
        "tags": [
          "Admin View Management"
        ],
        "operationId": "getUnCheckedVerifications",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationViewByAdminResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "UploadPassportRequest": {
        "type": "object",
        "properties": {
          "series": {
            "type": "string",
            "maxLength": 4,
            "minLength": 4
          },
          "number": {
            "type": "string",
            "maxLength": 6,
            "minLength": 6
          }
        },
        "required": [
          "number",
          "series"
        ]
      },
      "PassportVerificationResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "RegisterRequest": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 50,
            "minLength": 3
          },
          "email": {
            "type": "string",
            "minLength": 1
          },
          "password": {
            "type": "string",
            "maxLength": 2147483647,
            "minLength": 6
          },
          "phoneNumber": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          }
        }
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "AuthRequest": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "minLength": 1
          },
          "password": {
            "type": "string",
            "minLength": 1
          }
        }
      },
      "DocumentVerificationResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "submittedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "VerificationViewByAdminResponse": {
        "type": "object",
        "properties": {
          "numberOfPendingVerifications": {
            "type": "integer",
            "format": "int32"
          },
          "documents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DocumentVerificationResponse"
            }
          },
          "status": {
            "type": "string",
            "enum": [
              "NOT_VERIFIED",
              "IN_PROCESS",
              "VERIFIED",
              "CANCELLED"
            ]
          }
        }
      }
    }
  }
}